move group constants to its own file
