--- frontend/model/constants.js
@@ -1,5 +1,0 @@
-'use strict'
-
-// useful for 'okTurtles.data'
-// example (that we're no longer using):
-// export const WE_JUST_JOINED = '@const/just-joined'
