+++ frontend/model/contracts/constants.js
@@ -1,0 +1,11 @@
+'use strict'
+
+export const INVITE_INITIAL_CREATOR = 'INVITE_INITIAL_CREATOR'
+export const INVITE_STATUS = {
+  VALID: 'valid',
+  USED: 'used'
+}
+export const PROFILE_STATUS = {
+  ACTIVE: 'active',
+  REMOVED: 'removed'
+}
