+++ test/backend.test.js
@@ -14,7 +14,8 @@ import proposals from '~/frontend/model/contracts/voting/proposals.js'
 import { PROPOSAL_INVITE_MEMBER, PROPOSAL_REMOVE_MEMBER, PROPOSAL_GROUP_SETTING_CHANGE, PROPOSAL_PROPOSAL_SETTING_CHANGE, PROPOSAL_GENERIC } from '~/frontend/model/contracts/voting/constants.js'
 import { TYPE_MESSAGE } from '~/frontend/model/contracts/mailbox.js'
 import { PAYMENT_PENDING, PAYMENT_TYPE_MANUAL } from '~/frontend/model/contracts/payments/index.js'
-import { INVITE_INITIAL_CREATOR, createInvite } from '~/frontend/model/contracts/group.js'
+import { INVITE_INITIAL_CREATOR } from '~/frontend/model/contracts/constants.js'
+import { createInvite } from '~/frontend/model/contracts/group.js'
 // import '~/frontend/model/contracts/identity.js'
 import '~/frontend/model/state.js'
 import '~/frontend/controller/namespace.js'
